+++ src/main/java/com/supermartijn642/wirelesschargers/WirelessChargersClient.java
@@ -27,7 +27,7 @@ public class WirelessChargersClient {
             // Block entity renderers
             handler.registerCustomBlockEntityRenderer(type::getBlockEntityType, ChargerRenderer::new);
             // Baked item models
-            ResourceLocation location = new ModelResourceLocation("wirelesschargers:" + type.getRegistryName(), "inventory");
+            ResourceLocation location = new ModelResourceLocation(new ResourceLocation("wirelesschargers:" + type.getRegistryName()), "inventory");
             handler.registerModelOverwrite(location, CustomRendererBakedModelWrapper::wrap);
             // Item renderer
             handler.registerCustomItemRenderer(type::getItem, () -> new BlockEntityCustomItemRenderer<>(